0.0.2 update: changes data_save to data_processing module. Also changes structure of module to functional
